test(perps): harden Compression Streams polyfill coverage for Hermes
Guard the shim import, force the sync fflate branch (no workers), and
assert deflate-raw inflate against the polyfill rather than Node natives.
